Two green tests.
Bit of a faff with paths of all the config and aiml files. Have brought
in the AIMLbot project as part of the solution. That helped me debug the
file not found exceptions. Still a bit of work to do if we want to
deploy, but at least we have two green unit tests.